lien entre les fonctions
